drift-v2/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
redis==5.0.1
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
scipy==1.11.4
